### 2021-07-24 #3534 @whedon
[PRE REVIEW]: The Python Sky Model 3 software
**Submitting author:** @zonca (<a href="http://orcid.org/0000-0001-6841-1058">Andrea Zonca</a>)
**Repository:** <a href="https://github.com/galsci/pysm" target ="_blank">https://github.com/galsci/pysm</a>
**Version:** 3.3.1
