### 2021-07-30 #3534 @zonca
thanks @danielskatz
Can I post a pre-print of the paper to the Arxiv in the meantime?
