### 2021-07-30 #3534 @danielskatz
Yes, that's fine.  Also, be aware that you need to have an open source license - I see what I think is a BSD 3-clause license in a subdirectory, which JOSS did not find, so this is technically ok.  Normally, this would be in the top level of the repo and would be called LICENSE so that GitHub knows it is there and labels the repo correctly and JOSS knows it's there.  This placement is not required by JOSS, but is a common practice.
