+++ setup.cfg
@@ -3,7 +3,7 @@ name = pysm3
 author = relevant to CMB experimentAndrea Zonca
 author_email = code@andreazonca.com
 license = BSD 3-Clause
-license_file = licenses/LICENSE.rst
+license_file = LICENSE.rst
 url = https://pysm3.readthedocs.io/
 description = "PySM generates full-sky simulations of Galactic emissions in intensity and polarization"
 long_description = file: README.rst
