### 2021-08-01 9c91011
get also tex files in github action
See https://github.com/openjournals/joss/issues/132
