+++ .github/workflows/paper.yaml
@@ -13,17 +13,15 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@master
+      - name: TeX and PDF
+        uses: docker://openjournals/paperdraft:latest
         with:
-          journal: joss
-          # This should be the path to the paper within your repo.
-          paper-path: paper/paper.md
+          args: '-k paper/paper.md'
+        env:
+          GIT_SHA: $GITHUB_SHA
+          JOURNAL: joss
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
           name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
-          path: paper/paper.pdf
+          path: paper/
