### 2021-08-04 #3534 @zonca
thanks @danielskatz, I moved the license to the root of the repository, I don't know how to tell whedon to check again.

I also published the preprint to Arxiv: https://arxiv.org/abs/2108.01444
