### 2021-08-04 52e31b3
fix typo in JOSS badge
