### 2021-08-04 5c339f0
Update citation year in README
Original paper accepted in 2017.
