### 2021-08-16 #3534 @zonca
thanks @arfon!

first choice would be cosmology + python, specific CMB knowledge first: Pqrs6, sibirrer, smsharma, benjaminrose, danehkar, yymao, aureliocarnero, changhoonhahn, Christovis, rmorgan10, villaa
