### 2021-09-19 #3534 @arfon
:wave: @aureliocarnero @Pqrs6 @sibirrer – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration here is [_The Python Sky Model 3 software_](https://github.com/galsci/pysm).
