### 2021-09-25 d6bf7da
Added notebook for sync template work
