+++ sync_poltens.py
@@ -29,14 +29,15 @@ import pysm3.units as u
 import cosmoplotian.colormaps
 import matplotlib.pyplot as plt
 import matplotlib as mpl
+
+CMAP = mpl.cm.get_cmap("div yel grn")
 plt.rcParams['text.usetex'] = True
-cmap = mpl.cm.get_cmap("div yel grn")
-mpl.rcParams['axes.prop_cycle'] = mpl.cycler(color=["k"] + [cmap(i) for i in [0.2, 0.5, 0.8]]) 
+plt.rcParams['axes.prop_cycle'] = mpl.cycler(color=["k"] + [CMAP(i) for i in [0.2, 0.5, 0.8]]) 
+#plt.rcParams['savefig.directory'] = Path("./plots")
+plt.rcParams['image.cmap'] = "div yel grn"
 
 FLAGS = flags.FLAGS
-
 DATA_DIR = Path(os.environ["PYSM_DATA_DIR"])
-
 MAPS = {
     'PolMask': "COM_Mask_CMB-common-Mask-Pol_2048_R3.00.fits",
     'IntMask': "COM_Mask_CMB-common-Mask-Int_2048_R3.00.fits",
@@ -142,35 +143,10 @@ def DoFrolovTransform():
 
     # 0.88 degrees Kband res
     # 56 arcmin beam size haslam
-
-    #IQU = hp.smoothing(IQU, fwhm=1. * np.pi/180.)
-
     iqu = FrolovTransformForward(IQU)
-
-    hp.mollview(iqu[0], min=-2, max=10, title=r"Frolov $i$")
-    fig = plt.gcf()
-    fig.savefig("transformed_i_naive.pdf")
-    hp.mollview(iqu[1], min=-5, max=5, title=r"Frolov $q$")
-    fig = plt.gcf()
-    fig.savefig("transformed_q_naive.pdf")
-    hp.mollview(iqu[2], min=-5, max=5, title=r"Frolov $u$")
-    fig = plt.gcf()
-    fig.savefig("transformed_u_naive.pdf")
-
     hp.write_map(str(DATA_DIR / "transformed_iqu_naive.fits"), iqu, overwrite=True)
 
     iqu = FrolovTransformForward(hp.smoothing(IQU, fwhm=np.pi/180.))
-
-    hp.mollview(iqu[0], min=-2, max=10, title=r"${\rm Frolov}~i$")
-    fig = plt.gcf()
-    fig.savefig("transformed_i_2degfwhm.pdf")
-    hp.mollview(iqu[1], min=-2, max=2, title=r"${\rm Frolov}~q$")
-    fig = plt.gcf()
-    fig.savefig("transformed_q_2degfwhm.pdf")
-    hp.mollview(iqu[2], min=-2, max=2, title=r"${\rm Frolov}~u$")
-    fig = plt.gcf()
-    fig.savefig("transformed_u_2degfwhm.pdf")
-
     hp.write_map(str(DATA_DIR / "transformed_iqu_fwhm2deg.fits"), iqu, overwrite=True)
 
     IQU = hp.smoothing(IQU, fwhm=2 * np.pi/180.)
@@ -179,18 +155,62 @@ def DoFrolovTransform():
     rescaling = np.sqrt(0.99 * IQU[0] ** 2 / P ** 2)
     IQU[1:, idx] *= rescaling[None, idx]
     iqu = FrolovTransformForward(IQU)
+    hp.write_map(str(DATA_DIR / "transformed_iqu_fwhm2deg_corrected.fits"), iqu, overwrite=True)
+    return 
 
-    hp.mollview(iqu[0], min=-2, max=10, title=r"${\rm Frolov}~i$")
+
+def PlotFrolovTransformediqu():
[…]
