+++ pysm3/co_lines.py
@@ -1,0 +1,227 @@
+import numpy as np
+
+import healpy as hp
+
+try:  # PySM >= 3.2.1
+    import pysm3.units as u
+    import pysm3 as pysm
+except ImportError:
+    import pysm.units as u
+    import pysm
+
+from .utils import RemoteData
+
+
+class COLines(pysm.Model):
+    def __init__(
+        self,
+        target_nside,
+        output_units,
+        has_polarization=True,
+        line="10",
+        include_high_galactic_latitude_clouds=False,
+        polarization_fraction=0.001,
+        theta_high_galactic_latitude_deg=20.0,
+        random_seed=1234567,
+        verbose=False,
+        run_mcmole3d=False,
+        map_dist=None,
+        coord="C",
+    ):
+
+        """Class defining attributes for CO line emission.
+        CO templates are extracted from Type 1 CO Planck maps.
+        See further details in https://www.aanda.org/articles/aa/abs/2014/11/aa21553-13/aa21553-13.html
+
+        Parameters
+        ----------
+        target_nside : int
+            HEALPix NSIDE of the output maps
+        output_units : str
+            unit string as defined by `pysm.convert_units`, e.g. uK_RJ, K_CMB
+        has_polarization : bool
+            whether or not to simulate also polarization maps
+        line : string
+            CO rotational transitions.
+            Accepted values : 10, 21, 32
+        polarization_fraction: float
+            polarisation fraction for polarised CO emission.
+        include_high_galactic_latitude_clouds: bool
+            If True it includes a simulation from MCMole3D to include
+            high Galactic Latitude clouds.
+            (See more details at http://giuspugl.github.io/mcmole/index.html)
+        run_mcmole3d: bool
+            If True it simulates  HGL cluds by running MCMole3D, otherwise it coadds
+            a map of HGL emission.
+        random_seed: int
+            set random seed for mcmole3d simulations.
+        theta_high_galactic_latitude_deg : float
+            Angle in degree  to identify High Galactic Latitude clouds
+            (i.e. clouds whose latitude b is `|b|> theta_high_galactic_latitude_deg`).
+        map_dist : mpi4py communicator
+            Read inputs across a MPI communicator, see pysm.read_map
+        """
+
+        self.line = line
+        self.line_index = {"10": 0, "21": 1, "32": 2}[line]
+        self.line_frequency = {
+            "10": 115.271 * u.GHz,
+            "21": 230.538 * u.GHz,
+            "32": 345.796 * u.GHz,
+        }[line]
+        self.target_nside = target_nside
+
+        self.template_nside = 512 if self.target_nside <= 512 else 2048
+
+        super().__init__(nside=target_nside, map_dist=map_dist)
+
+        self.remote_data = RemoteData(coord)
+
+        self.planck_templatemap_filename = "co/HFI_CompMap_CO-Type1_{}_R2.00_ring.fits".format(
+            self.template_nside
+        )
+        self.planck_templatemap = self.read_map(
+            self.remote_data.get(self.planck_templatemap_filename),
+            field=self.line_index,
+            unit=u.K_CMB,
+        )
+
+        self.include_high_galactic_latitude_clouds = (
+            include_high_galactic_latitude_clouds
+        )
+        self.has_polarization = has_polarization
+        self.polarization_fraction = polarization_fraction
+        self.theta_high_galactic_latitude_deg = theta_high_galactic_latitude_deg
+        self.random_seed = random_seed
+        self.run_mcmole3d = run_mcmole3d
+
+        self.output_units = u.Unit(output_units)
+        self.verbose = verbose
+
+    def signal(self):
+        """
+        Simulate CO signal
+        """
+        out = (
[…]
