+++ pysm3/__init__.py
@@ -22,3 +22,4 @@ from . import units
 from .distribution import MapDistribution
 from .mpi import mpi_smoothing
 from .utils import normalize_weights, bandpass_unit_conversion, check_freq_input
+from .co_lines import COLines
