### 2021-09-29 c652907
pysm3 doesn't support different coordinates for templates
