### 2021-09-29 fc11f29
fix: remove unused coord parameter
