### 2021-09-29 836cf1a
shorten lines to make flake8 happy
