### 2021-09-29 #3783 @whedon
[REVIEW]: The Python Sky Model 3 software
**Submitting author:** @zonca (<a href="http://orcid.org/0000-0001-6841-1058">Andrea Zonca</a>)
**Repository:** <a href="https://github.com/galsci/pysm" target ="_blank">https://github.com/galsci/pysm</a>
**Version:** v3.3.2
