### 2021-10-02 #3783 @patricialarsen
Hi!

Installation is going fairly smoothly so far. The conda, pip and development installation work from the instructions on my home linux computer, with a fresh python3 installation.

A few minor issues though once I hit running the unit tests on these - firstly there's additional notes on this in the setup.py file. Those are useful but a little well-hidden and could be placed somewhere more visible. Secondly a slight issue is coming up in the config file for an install of the development version into a fresh conda environment:

I'm getting
pytest: error: unrecognized arguments: --doctest-rst

It works okay if I change the flag in line 45 of setup.cfg from
addopts = --doctest-rst
to
addopts = --doctest-glob="*.rst"

I'm having a little more trouble getting the tests running through tox, which seem to be fixed by adding dependencies to the config file as well as updating my c compilers through
conda install -c conda-forge c-compiler compilers cxx-compiler

The dependencies I had to add to the deps section of tox.ini are:
    astropy
    pytest
    numba
    healpy

I then manage to pass tests using tox with
=================================== 62 passed, 3 skipped, 585 warnings in 55.26s ===================================
And through pytest with
=================================== 60 passed, 2 skipped, 340 warnings in 52.24s ===================================

A large number of warnings, which mostly look like
 UserWarning: Retrieve data for pysm_2_test_data/check_d5_100_uK_RJ_64.fits (if not cached already)

These warnings could be looked into, and suppressed if they're not important to avoid flagging as an issue.

@smsharma when you look at the installation/automated tests could you please keep an eye out for if these tests run straightforwardly on your system and if you need any adjustments to get them to run?
