### 2021-10-06 6e76285
feat: implemented dust layers model, test trivial case
