### 2021-10-06 af44512
doc: modify docstring
