### 2021-10-07 ec529cc
Remove doctest-rst option from pytest
We do not have any code block in the documentation
See
https://github.com/openjournals/joss-reviews/issues/3783#issuecomment-932672403
