+++ setup.cfg
@@ -42,7 +42,6 @@ testpaths = "pysm3" "docs"
 astropy_header = true
 doctest_plus = enabled
 text_file_format = rst
-addopts = --doctest-rst
 
 [coverage:run]
 omit =
