### 2021-10-07 f578230
Rewrite test_units.py
Use pytest instead of unittest, simplify using tmp_path
