### 2021-10-07 6ca3741
move also bandpass_unit_conversion to logging
