### 2021-10-07 aec18c7
fix: escape \e that was causing DeprecationWarning
