### 2021-10-07 f6eaaf0
fix: extract_hdu_unit was leaving FITS file open
