### 2021-10-07 d424dfd
fix: filter out benign numpy warnings
https://stackoverflow.com/questions/40845304/runtimewarning-numpy-dtype-size-changed-may-indicate-binary-incompatibility
