### 2021-10-07 3053e5f
Filter out AstropyDeprecationWarning
