### 2021-10-07 79f4555
filter out all numpy benign warnings
