### 2021-10-07 917520b
fix: numpy warning, use float32 instead of float
