### 2021-10-07 23e875f
fix: use filterwarnings instead of simplefilter
