+++ pysm3/__init__.py
@@ -4,7 +4,7 @@ from ._astropy_init import *  # noqa
 import warnings
 from astropy.utils.exceptions import AstropyDeprecationWarning
 
-warnings.simplefilter("ignore", category=AstropyDeprecationWarning)
+warnings.filterwarnings("ignore", category=AstropyDeprecationWarning)
 warnings.filterwarnings("ignore", message="may indicate binary incompatibility")
 
 try:
