### 2021-10-07 137c49d
fix: remove verbose which is deprecated in healpy 1.15
