### 2021-10-07 586dded
fix: no need of deprecation warning filter anymore
