### 2021-10-07 e1acb81
mention warnings to logging and JOSS review in changelog
