+++ pysm3/data/presets.cfg
@@ -137,6 +137,15 @@ nside_uval = 256
 seed = 4632
 f_car = 1.0
 f_fe = 0.44
+[d12]
+class = "ModifiedBlackBodyLayers"
+map_layers = "mkd_dust/thermaldust_ampl{layer}.fits"
+freq_ref = "353 GHz"
+map_mbb_index = "mkd_dust/thermaldust_specind{layer}.fits"
+map_mbb_temperature = "mkd_dust/thermaldust_temp{layer}.fits"
+num_layers = 6
+unit_layers = "MJy/sr"
+unit_mbb_temperature = "K"
 [s0]
 class = "PowerLaw"
 map_I = "pysm_2/synch_t_new.fits"
