### 2021-10-13 4f26234
fix: unit was not set in input layers, test with fixtures
