### 2021-10-15 d5306f5
generate lower nside templates script
