### 2021-10-15 4fbe691
add metadata script
