### 2021-10-27 c57c3e5
[doc] how to install test requirements
From:
https://github.com/openjournals/joss-reviews/issues/3783#issuecomment-950407911
