+++ setup.cfg
@@ -26,7 +26,8 @@ install_requires =
 [options.extras_require]
 test =
     pytest-astropy
-    mpi4py
+    pytest
+    tox
 docs =
     sphinx-astropy
     nbsphinx
