### 2021-10-27 9bb6228
[docs] mention to install pandoc
