### 2021-10-27 c2d470f
[rtd] install test requirements
