+++ .readthedocs.yml
@@ -11,6 +11,6 @@ python:
       path: .
       extra_requirements:
         - docs
-        - all
+        - test
 
 formats: []
