### 2021-10-27 4847b85
[rtd] switch to python 3.8
