+++ .readthedocs.yml
@@ -4,7 +4,7 @@ build:
   image: latest
 
 python:
-  version: 3.7
+  version: 3.8
   system_packages: true
   install:
     - method: pip
