### 2021-10-28 f55b97e
[docs] link to MPI tutorial
