### 2021-10-28 9ace3e6
[fix] setup mpi requirement
