+++ setup.cfg
@@ -28,6 +28,8 @@ test =
     pytest-astropy
     pytest
     tox
+mpi =
+    mpi4py
 docs =
     sphinx-astropy
     nbsphinx
