+++ tox.ini
@@ -63,6 +63,7 @@ deps =
 # The following indicates which extras_require from setup.cfg will be installed
 extras =
     test
+    mpi
     alldeps: all
 
 commands =
