### 2021-10-29 6345750
[fix] fix imports in so pysm model example
