### 2021-10-29 c4a1627
[docs] improve CMBMap
