### 2021-10-29 #3783 @christinahedges
@zonca it seems like @patricialarsen and @smsharma have both agreed that your submission should be accepted! This is great news. I believe there are a few unchecked marks in the review, but that your new merged PR will solve these unchecked marks. Reading the thread, I believe the automated tests checkbox from @patricialarsen is unchecked only because of the warnings that occurred during execution.

It looks like you have no broken references in the paper. I find no significant copy editing necessary in your proof. I would suggest in the "Performance" section you might add a sentence explaining that the level to which the results of the PySM 2 and PySM3 models at Nside = 512 are identical, since I believe they should ideally produce close to identical results. This is only a gentle suggestion.

@zonca if you are happy and would like to make no further changes based on the review, please make a tagged release 🏷️  of the code (since the code and documentation has changed during the review) and archive it, and then please report the version number and archive DOI here in this thread. Please make sure this archived version has all the correct meta data (author names, title etc)!
