### 2021-10-29 #3783 @smsharma
Thanks @zonca for taking the comments and suggestions into account, with their changes I'm happy to check off on the remaining items as well @christinahedges . Congrats to the authors on a very useful package!
