### 2021-10-29 87d8f14
[paper] mention tolerance in unit tests
