### 2021-10-29 #3783 @zonca
thanks @christinahedges and @smsharma, you reported a lot of issues that nobody bothers to report so they never get fixed and make life for new users a lot worse! The review was very useful, thanks @patricialarsen.

Ok, I released 3.3.2
https://zenodo.org/record/5620869
