### 2021-11-01 #3783 @danielskatz
In addition, I've suggested some changes to the paper in https://github.com/galsci/pysm/pull/94

And finally, many of your references don't have journals - this is because you are using astro-standards abbreviations for journals in some cases, which JOSS doesn't support.  Please change things like `journal = {\apj}` to spell out the journal names instead.
