### 2021-11-01 #3783 @zonca
> @zonca - Please make sure this archived version has all the correct meta data (author names, title etc)! You can change zenodo metadata without changing the DOI...

thanks @danielskatz, yes, the title is just the package name with the version, authors are fine.
