### 2021-11-01 #3783 @danielskatz
> thanks @danielskatz, yes, the title is just the package name with the version, authors are fine.

yes, I realize this, and I am asking you to change the title of the archive in Zenodo to match the title of the paper
