### 2021-11-01 #3783 @danielskatz
@zonca - sorry, one more thing I missed - can you protect the `P`s in Python in bib entries with `{P}` - in Thorne et al and Zonca et al

otherwise, this is ready to publish
