### 2021-11-01 61363ab
[paper] fix bibliography capitalization
