+++ paper/paper.bib
@@ -24,12 +24,12 @@
   number = {35},
   pages = {1298},
   author = {Andrea Zonca and Leo P. Singer and Daniel Lenz and Martin Reinecke and Cyrille Rosset and Eric Hivon and Krzysztof M. Gorski},
-  title = {healpy: equal area pixelization and spherical harmonics transforms for data on the sphere in Python},
+  title = {{healpy: equal area pixelization and spherical harmonics transforms for data on the sphere in Python}},
   journal = {Journal of Open Source Software}
 }
 
 @article{pysm17,
-   title={The Python Sky Model: software for simulating the Galactic microwave sky},
+   title={{The Python Sky Model: software for simulating the Galactic microwave sky}},
    volume={469},
    ISSN={1365-2966},
    url={http://dx.doi.org/10.1093/mnras/stx949},
