### 2021-11-01 #3783 @zonca
@danielskatz never noticed before that double braces `{{}}` preserves capitalization... ;)
